+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Fix `getDotPath` function for empty issue path
+
 ## v0.1.0 (October 06, 2024)
 
 - Initial release